add del method in meta

+++ stochos-back/src/main/java/br/com/tcc/stochos/controller/GrupoController.java
@@ -41,8 +41,14 @@ public class GrupoController {
   }
 
   @DeleteMapping("/{id}")
-  public HttpStatus deletarGrupo(@PathVariable("id") Long id){
-     grupoRepository.deleteById(id);
-     return HttpStatus.ACCEPTED;
+  public ResponseEntity deletarGrupo(@PathVariable("id") Long id){
+    {
+      if (grupoRepository.existsById(id)){
+        grupoRepository.deleteById(id);
+        return new ResponseEntity<>(HttpStatus.ACCEPTED);
+
+      }
+      return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
   }
 }
